pyproject.toml
[project]
name = "nxtint"
version = "0.1.0"
description = "Integer sequence prediction using transformers"
authors = [
    { name = "Ivo Maceira", email = "ivomaceira@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
	"python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
	"ipython>=8.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
nxtint = "nxtint.__main__:main"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "D", "I", "N", "S", "T", "UP", "B", "SIM", "R"]
ignore = ["B019", "D413", "S101", "S603", "S607", "T201"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=nxtint --cov-report=term-missing"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(levelname)s - %(module)s - %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


